Adds simple test-report template
